+++ lib/msf/core/exploit/remote/ldap/server.rb
@@ -9,9 +9,7 @@ module Msf
   module Exploit::Remote::LDAP
     module Server
       include Exploit::Remote::SocketServer
-      include Rex::Proto::LDAP::Server::Auth
 
-      #
       # Initializes an exploit module that serves LDAP requests
       #
       def initialize(info = {})
@@ -72,6 +70,19 @@ module Msf
         cli.write(data)
       end
 
+      # Perform the required tasks after pdu object is processed
+      # Override this method in modules to take flow control
+      #
+      def on_processed_pdu(processed_data)
+        pdu_type = processed_data[:pdu_type]
+        case pdu_type
+        when Net::LDAP::PDU::BindRequest
+          processed_data
+        else
+          return
+        end
+      end
+
       #
       # Starts the server
       #
@@ -79,15 +90,25 @@ module Msf
         comm = _determine_server_comm(bindhost)
         self.service = Rex::ServiceManager.start(
           Rex::Proto::LDAP::Server,
-          bindhost,
-          bindport,
-          datastore['LdapServerUdp'],
-          datastore['LdapServerTcp'],
-          read_ldif,
-          comm,
+          {
+            bindhost: bindhost,
+            bindport: bindport,
+            udp: datastore['LdapServerUdp'],
+            tcp: datastore['LdapServerTcp'],
+            ldif: read_ldif,
+            comm: comm,
+            challenge: datastore['CHALLENGE'],
+            domain: datastore['Domain'],
+            server: datastore['Server'],
+            dnsname: datastore['DnsName'],
+            dnsdomain: datastore['DnsDomain']
+          },
           { 'Msf' => framework, 'MsfExploit' => self }
         )
 
+        service.processed_pdu_handler do |processed_data|
+          on_processed_pdu(processed_data)
+        end
         service.dispatch_request_proc = proc do |cli, data|
           on_dispatch_request(cli, data)
         end
