+++ lib/rex/proto/ldap/server.rb
@@ -7,7 +7,7 @@ module Rex
   module Proto
     module LDAP
       class Server
-        attr_reader :serve_udp, :serve_tcp, :sock_options, :udp_sock, :tcp_sock, :syntax, :ldif
+        attr_reader :serve_udp, :serve_tcp, :sock_options, :udp_sock, :tcp_sock, :syntax, :ldif, :auth_options, :pdu_process
 
         module LdapClient
           attr_accessor :authenticated
@@ -51,29 +51,32 @@ module Rex
         #
         # Create LDAP Server
         #
-        # @param lhost [String] Listener address
-        # @param lport [Fixnum] Listener port
-        # @param udp [TrueClass, FalseClass] Listen on UDP socket
-        # @param tcp [TrueClass, FalseClass] Listen on TCP socket
-        # @param ldif [String] LDIF data
+        # @param opts [Hash] Options required to start service
         # @param ctx [Hash] Framework context for sockets
         # @param dblock [Proc] Handler for :dispatch_request flow control interception
         # @param sblock [Proc] Handler for :send_response flow control interception
         #
         # @return [Rex::Proto::LDAP::Server] LDAP Server object
-        def initialize(lhost = '0.0.0.0', lport = 389, udp = true, tcp = true, ldif = nil, comm = nil, ctx = {}, dblock = nil, sblock = nil)
-          @serve_udp = udp
-          @serve_tcp = tcp
+        def initialize(opts = {}, ctx = {}, dblock = nil, sblock = nil)
+          @serve_udp = opts[:udp]
+          @serve_tcp = opts[:tcp]
           @sock_options = {
-            'LocalHost' => lhost,
-            'LocalPort' => lport,
+            'LocalHost' => opts[:bindhost],
+            'LocalPort' => opts[:bindport],
             'Context' => ctx,
-            'Comm' => comm
+            'Comm' => opts[:comm]
           }
-          @ldif = ldif
+          @ldif = opts[:ldif]
           self.listener_thread = nil
           self.dispatch_request_proc = dblock
           self.send_response_proc = sblock
+          @auth_options  = { #check for nill cases
+            challenge: opts[:challenge],
+            domain: opts[:domain],
+            server: opts[:server],
+            dnsname: opts[:dnsname],
+            dnsdomain: opts[:dnsdomain]
+          }
         end
 
         #
@@ -114,6 +117,10 @@ module Rex
             end
           end
 
+          if auth_options
+            @auth_provider = Rex::Proto::LDAP::Auth.new(auth_options)
+          end
+
           self
         end
 
@@ -149,76 +156,126 @@ module Rex
         #
         # Default LDAP request dispatcher
         #
-        # @param cli [Rex::Socket::Tcp, Rex::Socket::Udp] Client sending the request
+        # @param client [Rex::Socket::Tcp, Rex::Socket::Udp] Client sending the request
         # @param data [String] raw LDAP request data
-        def default_dispatch_request(cli, data)
-          return if data.strip.empty?
+        def default_dispatch_request(client, data)
+          return if data.strip.empty? || data.strip.nil?
+
+          state = {}
+          post_pdu = false
+          processed_pdu_data = {}
+          state[client] = {
+            name: "#{client.peerhost}:#{client.peerport}",
+            ip: client.peerhost,
+            port: client.peerport,
+            service_name: 'ldap'
+          }
 
           data.extend(Net::BER::Extensions::String)
           begin
             pdu = Net::LDAP::PDU.new(data.read_ber!(Net::LDAP::AsnSyntax))
-            wlog("LDAP request has remaining data: #{data}") if !data.empty?
-            resp = case pdu.app_tag
-                   when Net::LDAP::PDU::BindRequest # bind request
-                     cli.authenticated = true
-                     encode_ldap_response(
-                       pdu.message_id,
-                       Net::LDAP::ResultCodeSuccess,
-                       '',
-                       '',
-                       Net::LDAP::PDU::BindResult
-                     )
-                   when Net::LDAP::PDU::SearchRequest # search request
-                     if cli.authenticated
-                       # Perform query against some loaded LDIF structure
-                       treebase = pdu.search_parameters[:base_object].to_s
-                       # ... search, build packet, send to client
-                       encode_ldap_response(
-                         pdu.message_id,
-                         Net::LDAP::ResultCodeNoSuchObject, '',
-                         Net::LDAP::ResultStrings[Net::LDAP::ResultCodeNoSuchObject],
-                         Net::LDAP::PDU::SearchResult
-                       )
-                     else
-                       service.encode_ldap_response(pdu.message_id, 50, '', 'Not authenticated', Net::LDAP::PDU::SearchResult)
-                     end
-                   when Net::LDAP::PDU::UnbindRequest
-                     nil # close client, no response can be sent over unbound comm
-                   else
-                     service.encode_ldap_response(
-                       pdu.message_id,
-                       Net::LDAP::ResultCodeUnwillingToPerform,
-                       '',
-                       Net::LDAP::ResultStrings[Net::LDAP::ResultCodeUnwillingToPerform],
-                       Net::LDAP::PDU::SearchResult
-                     ) end
-            resp.nil? ? cli.close : send_response(cli, resp)
+            vprint_status("LDAP request data remaining: #{data}") unless data.empty?
+
+            res = case pdu.app_tag
+                  when Net::LDAP::PDU::BindRequest
+                    user_login = pdu.bind_parameters
+                    server_creds = ''
+                    processed_pdu_data = @auth_provider.process_login_request(user_login)
+                    processed_pdu_data = processed_pdu_data.merge(state[client])
+                    if processed_pdu_data[:result_code] == Net::LDAP::ResultCodeSaslBindInProgress
+                      server_creds = processed_pdu_data[:server_creds]
+                    else
+                      processed_pdu_data[:result_message] = "LDAP Login Attempt => From:#{processed_pdu_data[:name]}\n Username:#{processed_pdu_data[:user]}\n #{processed_pdu_data[:private_type]}:#{processed_pdu_data[:private]}\n"
+                      processed_pdu_data[:result_message] += " Domain:#{processed_pdu_data[:domain]}" if processed_pdu_data[:domain]
+                      post_pdu = true
+                    end
+                    processed_pdu_data[:pdu_type] = pdu.app_tag
+                    encode_ldap_response(
+                      pdu.message_id,
+                      processed_pdu_data[:result_code],
+                      '',
+                      Net::LDAP::ResultStrings[processed_pdu_data[:result_code]],
+                      Net::LDAP::PDU::BindResult,
+                      server_creds
+                    )
+                  when Net::LDAP::PDU::SearchRequest
+                    filter = Net::LDAP::Filter.parse_ldap_filter(pdu.search_parameters[:filter])
+                    attrs = pdu.search_parameters[:attributes].empty? ? :all : pdu.search_parameters[:attributes]
+                    res = search_result(filter, pdu.message_id, attrs)
+                    if res.nil? || res.empty?
+                      result_code = Net::LDAP::ResultCodeNoSuchObject
+                    else
+                      client.write(res)
+                      result_code = Net::LDAP::ResultCodeSuccess
+                    end
+                    processed_pdu_data[:pdu_type] = pdu.app_tag
+                    encode_ldap_response(
+                      pdu.message_id,
+                      result_code,
+                      '',
+                      Net::LDAP::ResultStrings[result_code],
+                      Net::LDAP::PDU::SearchResult
+                    )
+                  when Net::LDAP::PDU::UnbindRequest
+                    client.close
+                    nil
+                  else
+                    if suitable_response(pdu.app_tag)
+                      result_code = Net::LDAP::ResultCodeUnwillingToPerform
+                      encode_ldap_response(
+                        pdu.message_id,
+                        result_code,
+                        '',
+                        Net::LDAP::ResultStrings[result_code],
+                        suitable_response(pdu.app_tag)
+                      )
+                    else
+                      client.close
+                    end
+                  end
+
+            if @pdu_process && post_pdu
+              @pdu_process.call(processed_pdu_data)
+            end
+            send_response(client, res) unless res.nil?
           rescue StandardError => e
-            elog(e)
-            cli.close
-            raise e
+            client.close
           end
         end
 
         #
         # Encode response for LDAP client consumption
         #
-        # @param msgid [Integer] LDAP message identifier
-        # @param code  [Integer] LDAP message code
-        # @param dn    [String]  LDAP distinguished name
-        # @param msg   [String]  LDAP response message
-        # @param tag   [Integer] LDAP response tag
+        # @param msgid        [Integer] LDAP message identifier
+        # @param code         [Integer] LDAP message code
+        # @param dn           [String]  LDAP distinguished name
+        # @param msg          [String]  LDAP response message
+        # @param tag          [Integer] LDAP response tag
+        # @param server_creds [String]  LDAP response server SASL credentials
         #
         # @return [Net::BER::BerIdentifiedOid] LDAP query response
-        def encode_ldap_response(msgid, code, dn, msg, tag)
-          [
-            msgid.to_ber,
+        def encode_ldap_response(msgid, code, dn, msg, tag, server_creds = '')
+          case code
+          when Net::LDAP::ResultCodeSaslBindInProgress
             [
-              code.to_ber_enumerated,
-              dn.to_ber,
-              msg.to_ber
-            ].to_ber_appsequence(tag)
-          ].to_ber_sequence
+              msgid.to_ber,
+              [
+                code.to_ber_enumerated,
+                dn.to_ber,
+                msg.to_ber,
+                [server_creds.to_ber].to_ber_contextspecific(7)
+              ].to_ber_appsequence(tag)
+            ].to_ber_sequence
+          else
+            [
+              msgid.to_ber,
+              [
+                code.to_ber_enumerated,
+                dn.to_ber,
+                msg.to_ber
+              ].to_ber_appsequence(tag)
+            ].to_ber_sequence
+          end
         end
 
         #
@@ -228,25 +285,50 @@ module Rex
         # @param attrflt [Array, Symbol] LDAP attribute filter
         #
         # @return [Array] Query matches
-        def search_ldif(filter, msgid, attrflt = :all)
-          return [] if @ldif.nil? || @ldif.empty?
 
-          ldif.map do |dn, entry|
-            next unless filter.match(entry)
-
+        def search_result(filter, msgid, attrflt = :all)
+          if @ldif.nil? || @ldif.empty?
             attrs = []
-            entry.each do |k, v|
-              if attrflt == :all || attrflt.include?(k.downcase)
-                attrvals = v.map(&:to_ber).to_ber_set
-                attrs << [k.to_ber, attrvals].to_ber_sequence
+            if attrflt.is_a?(Array)
+              attrflt.each do |at|
+                attrval = [Rex::Text.rand_text_alphanumeric(10)].map(&:to_ber).to_ber_set
+                attrs << [at.to_ber, attrval].to_ber_sequence
               end
+              dn = "dc=#{Rex::Text.rand_text_alphanumeric(10)},dc=#{Rex::Text.rand_text_alpha(4)}"
+              appseq = [
+                dn.to_ber,
+                attrs.to_ber_sequence
+              ].to_ber_appsequence(Net::LDAP::PDU::SearchReturnedData)
+              [msgid.to_ber, appseq].to_ber_sequence
             end
-            appseq = [
-              dn.to_ber,
-              attrs.to_ber_sequence
-            ].to_ber_appsequence(Net::LDAP::PDU::SearchReturnedData)
-            [msgid.to_ber, appseq].to_ber_sequence
-          end.compact
+          else
+            ldif.map do |bind_dn, entry|
+              next unless filter.match(entry)
+
+              attrs = []
+              entry.each do |k, v|
+                if attrflt == :all || attrflt.include?(k.downcase)
+                  attrvals = v.map(&:to_ber).to_ber_set
+                  attrs << [k.to_ber, attrvals].to_ber_sequence
+                end
+              end
+              appseq = [
+                bind_dn.to_ber,
+                attrs.to_ber_sequence
+              ].to_ber_appsequence(Net::LDAP::PDU::SearchReturnedData)
+              [msgid.to_ber, appseq].to_ber_sequence
+            end.compact.join
+          end
+        end
+
+        #
+        # Sets the tasks to be performed after processing of pdu object
+        #
+        # @param proc [Proc] block of code to execute
+        #
+        # @return pdu_process [Proc] steps to be executed
+        def processed_pdu_handler(&proc)
+          @pdu_process = proc if block_given?
         end
 
         #
@@ -254,6 +336,27 @@ module Rex
         #
         def self.hardcore_alias(*args)
           "#{args[0] || ''}-#{args[1] || ''}-#{args[4] || ''}"
+        end
+
+        #
+        # Get suitable response for a particular request
+        #
+        # @param request [Integer] Type of request
+        #
+        # @return response [Integer] Type of response
+        def suitable_response(request)
+          responses = {
+            Net::LDAP::PDU::BindRequest => Net::LDAP::PDU::BindResult,
+            Net::LDAP::PDU::SearchRequest => Net::LDAP::PDU::SearchResult,
+            Net::LDAP::PDU::ModifyRequest => Net::LDAP::PDU::ModifyResponse,
+            Net::LDAP::PDU::AddRequest => Net::LDAP::PDU::AddResponse,
+            Net::LDAP::PDU::DeleteRequest => Net::LDAP::PDU::DeleteResponse,
+            Net::LDAP::PDU::ModifyRDNRequest => Net::LDAP::PDU::ModifyRDNResponse,
+            Net::LDAP::PDU::CompareRequest => Net::LDAP::PDU::CompareResponse,
+            Net::LDAP::PDU::ExtendedRequest => Net::LDAP::PDU::ExtendedResponse
+          }
+
+          responses[request]
         end
 
         #
